+++ src/Kentico.Kontent.ModelGenerator.Core/ManagementCodeGenerator.cs
@@ -42,20 +42,7 @@ namespace Kentico.Kontent.ModelGenerator.Core
             return 0;
         }
 
-        internal async Task GenerateContentTypeModels()
-        {
-            var classCodeGenerators = await GetClassCodeGenerators();
-
-            if (!classCodeGenerators.Any())
-            {
-                Console.WriteLine(NoContentTypeAvailableMessage);
-                return;
-            }
-
-            WriteToOutputProvider(classCodeGenerators);
-        }
-
-        internal async Task<ICollection<ClassCodeGenerator>> GetClassCodeGenerators()
+        internal override async Task<ICollection<ClassCodeGenerator>> GetClassCodeGenerators()
         {
             var managementTypes = await GetAllContentModelsAsync(await _managementClient.ListContentTypesAsync());
             var managementSnippets = await GetAllContentModelsAsync(await _managementClient.ListContentTypeSnippetsAsync());
@@ -132,30 +119,6 @@ namespace Kentico.Kontent.ModelGenerator.Core
             var classFilename = $"{classDefinition.ClassName}";
 
             return ClassCodeGeneratorFactory.CreateClassCodeGenerator(Options, classDefinition, classFilename, true);
-        }
-
-        internal async Task GenerateBaseClass()
-        {
-            var classCodeGenerators = await GetClassCodeGenerators();
-
-            if (!classCodeGenerators.Any())
-            {
-                Console.WriteLine(NoContentTypeAvailableMessage);
-                return;
-            }
-
-            var baseClassCodeGenerator = new BaseClassCodeGenerator(Options.BaseClass, Options.Namespace);
-
-            foreach (var codeGenerator in classCodeGenerators)
-            {
-                baseClassCodeGenerator.AddClassNameToExtend(codeGenerator.ClassDefinition.ClassName);
-            }
-
-            var baseClassCode = baseClassCodeGenerator.GenerateBaseClassCode();
-            WriteToOutputProvider(baseClassCode, Options.BaseClass, false);
-
-            var baseClassExtenderCode = baseClassCodeGenerator.GenereateExtenderCode();
-            WriteToOutputProvider(baseClassExtenderCode, baseClassCodeGenerator.ExtenderClassName, true);
         }
 
         private static void AddProperty(ElementMetadataBase element, ref ClassDefinition classDefinition)
